refactor: relocate ritual models to core and update entity defaults
This commit moves fashion and beauty domain models from the KoColor-specific model package to a centralized core ritual module. It also refactors several Room entities to provide default values for timestamps and optional fields, improving consistency and reducing boilerplate during instantiation.

- **Model Relocation**:
    - Updated `FashionConverters`, `ClothingDefaults`, `CosmeticDefaults`, and all Room entities to import models (e.g., `ClothingCategory`, `SeasonalType`, `FashionAdvice`) from `com.zoewave.probase.core.model.ritual` instead of the local KoColor model package.
- **Entity Improvements**:
    - **`ClothingItemEntity.kt` & `CosmeticItemEntity.kt`**: Added `System.currentTimeMillis()` as the default value for the `timestamp` field.
    - **`FashionProfileEntity.kt`**: Added default `null` values for physical attributes (`skinToneHex`, `eyeColor`, `hairColor`, `notes`) and initialized `recommendedPalette` as an empty list.
    - **`SavedSuggestionEntity.kt`**: Cleaned up unused `kotlinx.datetime` imports.
- **Documentation**:
    - Added descriptive comments to `ClothingItemEntity` and `CosmeticItemEntity` to categorize metadata, usage, and logistics fields.
