+++ applications/kocolor/db/src/main/java/com/zoewave/probase/kocolor/db/entity/RoutineEntity.kt
@@ -2,8 +2,8 @@ package com.zoewave.probase.kocolor.db.entity
 
 import androidx.room3.Entity
 import androidx.room3.PrimaryKey
-import com.zoewave.probase.kocolor.model.RoutineStep
-import com.zoewave.probase.kocolor.model.RoutineTime
+import com.zoewave.probase.core.model.ritual.RoutineStep
+import com.zoewave.probase.core.model.ritual.RoutineTime
 
 @Entity(tableName = "beauty_routines")
 data class RoutineEntity(
